Agregamos archivos al manejo de modulos con JS
